requirements.txt
datasets
faiss-cpu
sentence-transformers
transformers
nltk
flask
requests
dotenv
lancedb
torch